requirement.txt
fastapi>=0.115.0
uvicorn>=0.30.0
groq>=0.13.0
streamlit>=1.40.0
httpx>=0.27.0
gspread>=6.1.0
google-auth>=2.35.0
